# app/Swagger/Schemas/ActivitySchema.php
<?php

namespace App\Swagger\Schemas;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'Activity',
    title: 'Activity',
    description: 'Сущность активности',
    properties: [
        new OA\Property(property: 'id', type: 'integer', example: 1),
        new OA\Property(property: 'name', type: 'string', example: 'Молочная продукция'),
        new OA\Property(property: 'parent_id', type: 'integer', example: null, nullable: true),
        new OA\Property(property: 'created_at', type: 'string', format: 'date-time'),
        new OA\Property(property: 'updated_at', type: 'string', format: 'date-time'),
    ]
)]
class ActivitySchema
{
}