# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API v1 Documentation",
        "description": "Документация для API версии 1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Основной сервер API v1"
        }
    ],
    "paths": {
        "/api/v1/activities": {
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Создать новую деятельность",
                "operationId": "a4bcc37a677c42ef4c818699823ab5d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Молочная продукция"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Деятельность успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Activity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or nesting level exceeded"
                    },
                    "404": {
                        "description": "Parent activity not found"
                    }
                }
            }
        },
        "/api/v1/locations/nearby": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Получить организации и здания в пределах заданного радиуса от точки",
                "operationId": "150e2c6781e38113292ae1e10183c0e1",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широка центральной точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота центральной точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус в километрах",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций и зданий в пределах радиуса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "organizations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Organization"
                                                    }
                                                },
                                                "buildings": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Building"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Radius is required or invalid coordinates"
                    }
                }
            }
        },
        "/api/v1/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "cf108afa69aa8de272788f058d3a478b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Organization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                }
            }
        },
        "/api/v1/organizations/building/{id}/": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по ID здания",
                "operationId": "8476bef268e36b10efb63495c2d6d3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций в здании",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
        "/api/v1/organizations/activity/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по ID вида деятельности",
                "operationId": "c953f08800d9eb6a5a0375740a30e08b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций с указанным видом деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                }
            }
        },
        "/api/v1/organizations/activity-tree/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Находит организации, связанные с указанным видом деятельности и всеми его вложенными подкатегориями",
                "description": "Получить организации по ID вида деятельности и его подкатегориям",
                "operationId": "a4f63d36bfb369629e1180b8d5fcfc37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций, связанных с видом деятельности и его подкатегориями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                }
            }
        },
        "/api/v1/organizations/search": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по названию",
                "operationId": "974fcce5c131501f587a8846e2a36dd8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций, соответствующих поисковому запросу",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Name parameter is required"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "description": "Сущность активности",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Молочная продукция"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Building",
                "description": "Сущность здания",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1, офис 3"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double",
                        "example": 37.6173
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "description": "Сущность организации",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО \"Рога и Копыта\""
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "8-923-666-13-13"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Activities",
            "description": "Activities"
        },
        {
            "name": "Location",
            "description": "Location"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}